fix for iphone5s drawable pool is fixed at 3 move PrivateCaps into Shared
